feat: 강의 생성 Service 로직 구현 및 타입 오류 수정 (#10)

+++ src/main/java/com/cramit/domain/lecture/Lecture.java
@@ -15,7 +15,7 @@ public class Lecture {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int lectureId;
+    private Long lectureId;
 
     @Column(nullable = false)
     private Long memberId;
